django-service/requirements.txt
asgiref >= 3.4.1
Django >= 3.2.5
dj-rest-auth >= 2.1.9
django-allauth >= 0.47.0
django-cors-headers >= 3.7.0
django-rest-action-serializer >= 1.1.0
django-webpack-loader >= 1.1.0
djangorestframework >= 3.12.4
djangorestframework-simplejwt >= 4.7.2
drf-action-serializer >= 1.0.5
gunicorn >= 20.1.0
psycopg2-binary >= 2.9.1
python-dateutil >= 2.8.2
pytz >= 2021.1
requests >= 2.26.0
sqlparse >= 0.4.1
typing-extensions >= 4.0.1
